Add stopgap 22.04 LTS Vagrantfile
Current pre-provided Vagrantfile for Mini-NDN is severely out of
date. Ideal goal would be an integration with Github Actions (#5347)
to automatically build and deploy this periodically or a script
to run installer on a base image; however, both of these do not
solve the lack of a currently working solution.

Change-Id: I91d142e0345ca492df8c94b39f6f2322b947ce01